+++ code-types/TypeConverter.ts
@@ -9,35 +9,51 @@ module TypeConverter {
      * @param type the type to stringify
      * @param [typeConverter] an optional converter for the type names in the type
      * @param [includeNullability=true] whether to include type nullability (i.e. the '?' part of 'int?[]')
+     * @param [nullableSymbol="?"] the symbol for nullability (i.e. '?' for C# or ' | null' for TypeScript)
      */
-    export function typeToString(type: CodeAst.Type, typeConverter?: (typeName: string) => string, includeNullability: boolean = true): string {
+    export function typeToString(type: CodeAst.Type, typeConverter?: (typeName: string) => string, includeNullability = true, nullableSymbol = "?"): string {
         var dst: string[] = [];
-        _typeToString(type, dst, typeConverter, includeNullability);
+        _typeToString(type, dst, typeConverter, includeNullability, nullableSymbol);
         return dst.join("");
     }
 
 
-    function _typeToString(type: CodeAst.Type, dst: string[], typeConverter: (typeName: string) => string, includeNullability: boolean): void {
+    function _typeToString(type: CodeAst.Type, dst: string[], typeConverter: (typeName: string) => string, includeNullability: boolean, nullableSymbol: string): void {
         dst.push(typeConverter ? typeConverter(type.typeName) : type.typeName);
 
         var childs = type.genericParameters;
         if (childs && childs.length > 0) {
             dst.push("<");
             for (var i = 0, sizeM1 = childs.length - 1; i < sizeM1; i++) {
-                _typeToString(childs[i], dst, typeConverter, includeNullability);
+                _typeToString(childs[i], dst, typeConverter, includeNullability, nullableSymbol);
                 dst.push(", ");
             }
-            _typeToString(childs[sizeM1], dst, typeConverter, includeNullability);
+            _typeToString(childs[sizeM1], dst, typeConverter, includeNullability, nullableSymbol);
             dst.push(">");
         }
 
         if (type.nullable && includeNullability) {
-            dst.push("?");
+            dst.push(nullableSymbol);
         }
 
         if (type.arrayDimensions) {
+            var needsParens = matchAny(dst, (s) => s.indexOf("|") > -1 || s.indexOf("&") > -1);
+            if (needsParens) {
+                dst.unshift("(");
+                dst.push(")");
+            }
             dst.push(new Array(type.arrayDimensions + 1).join("[]"));
         }
+    }
+
+
+    function matchAny<T>(ary: T[], filter: (value: T, index: number) => boolean) {
+        for (var i = 0, size = ary.length; i < size; i++) {
+            if (filter(ary[i], i)) {
+                return true;
+            }
+        }
+        return false;
     }
 
 
@@ -204,11 +220,12 @@ module TypeConverter {
          * @param typeTemplate a Type, or string where the format must be 'typeName?[][]...' where typeName has no generic parameters, and the '?' (nullability) and '[][]...' (array dimensions) are optional
          * @param returnUnknownTypes
          * @param [includeNullability=false]
+         * @param [nullableSymbol="?"]
          */
-        static parseAndConvertTypeTemplate(typeTemplate: string | CodeAst.Type, returnUnknownTypes: boolean, includeNullability: boolean = false): string {
+        static parseAndConvertTypeTemplate(typeTemplate: string | CodeAst.Type, returnUnknownTypes: boolean, includeNullability: boolean = false, nullableSymbol = "?"): string {
             return (typeof typeTemplate === "string"
                 ? TypeScript.parseAndConvertTypeTemplateString(typeTemplate, returnUnknownTypes, includeNullability)
-                : typeToString(typeTemplate, (t) => TypeScript.convertSimpleType(t, returnUnknownTypes), includeNullability));
+                : typeToString(typeTemplate, (t) => TypeScript.convertSimpleType(t, returnUnknownTypes), includeNullability, nullableSymbol));
         }
 
 
@@ -218,13 +235,17 @@ module TypeConverter {
          * @param returnUnknownTypes
          * @param [includeNullability=false]
          */
-        static parseAndConvertTypeTemplateString(typeTemplate: string, returnUnknownTypes: boolean, includeNullability?: boolean): string {
+        static parseAndConvertTypeTemplateString(typeTemplate: string, returnUnknownTypes: boolean, includeNullability?: boolean, nullableSymbol = "?"): string {
             var typeInfo = TypeConverter.parseTypeTemplate(typeTemplate);
             var arrayCount = typeInfo.arrayDimensions;
 
             var tsType = TypeScript.convertSimpleType(typeInfo.typeName, returnUnknownTypes);
 
-            return tsType + (typeInfo.nullable && includeNullability ? "?" : "") + (arrayCount > 0 ? new Array(arrayCount + 1).join("[]") : "");
+            tsType += (typeInfo.nullable && includeNullability ? nullableSymbol : "");
+
+            var needsParens = arrayCount > 0 && (tsType.indexOf("|") > -1 || tsType.indexOf("&") > -1);
+            if (needsParens) tsType = "(" + tsType + ")";
+            return tsType + (arrayCount > 0 ? new Array(arrayCount + 1).join("[]") : "");
         }